Create test cookbook for integration testing.
The test cookbook current comprises of two recipes, the default recipe
and the audit mode controls recipe. The latter of the two leverages
Chef Audit Mode to run ServerSpec inline without having to use busser.
This may, or may not be the right approach to testing cookbooks later
on, but it seems to be the direction in which Chef is taking things.