Fixed Windows and latest release installs.
The Windows install would only allow for a single version to be
installed at a time. This now allows multiple versions to be installed.
Moreover, the latest and latest-esr values were broken for all
platforms as it would never upgrade the version past the initially
installed version, it will now upgrade the version automatically and
handle installation detection on Windows correctly.